fix(v3/linux): wire common events + dbus monitors on GTK4 default path (#5474)
The GTK4 default build's `linuxApp.run()` (the post-#5463 default in
alpha.93) never called `setupCommonEvents()`, `monitorThemeChanges()`,
or `monitorPowerEvents()`. As a result, listeners registered for
`events.Common.ApplicationStarted`, `Common.ThemeChanged`,
`Common.SystemWillSleep` and `Common.SystemDidWake` never fired, since
nothing forwarded the underlying `events.Linux.*` events to their common
counterparts. Reverting to alpha.92 worked because the gtk3 default
already wired these in `application_linux_gtk3.go` (renamed from the
previous `application_linux.go`).

Restore parity by calling all three from the gtk4 `run()`, and extract
the two dbus monitor helpers to a build-tag-shared
`application_linux_dbus.go` so both stacks use the same implementation.

Verified with the existing `pkg/application/internal/tests/services/*`
suite on lin-node1: the startup/shutdown tests hang under the pre-fix
GTK4 default (`Common.ApplicationStarted` never fires, listener never
calls `app.Quit()`) and pass after the fix on both default and
`-tags gtk3` builds.
